feat: do not have separate modal for name claiming
