+++ src/views/UserProfile.tsx
@@ -44,7 +44,6 @@ import { useAccountBalances } from '../hooks/useAccountBalances';
 import { useAddressByChainName, useChainName } from '../hooks/useChainName';
 
 import AccountPortfolio from '@/components/Account/AccountPortfolio';
-import ClaimChainNameModal from '@/components/modals/ClaimChainNameModal';
 import ProfileEditModal from '../components/modals/ProfileEditModal';
 import { CONFIG } from '../config';
 import { useModal } from '../hooks';
@@ -105,7 +104,6 @@ export default function UserProfile({
 
   const [profile, setProfile] = useState<any>(null);
   const [editOpen, setEditOpen] = useState(false);
-  const [claimChainNameOpen, setClaimChainNameOpen] = useState(false);
   const [editInitialSection, setEditInitialSection] = useState<'profile' | 'x'>('profile');
 
   // Get tab from URL search params, default to "feed"
@@ -614,27 +612,14 @@ export default function UserProfile({
     }
   };
 
-  const openClaimChainName = () => {
-    setEditOpen(false);
-    setClaimChainNameOpen(true);
-  };
-
   const profileModals = (
-    <>
-      <ProfileEditModal
-        open={editOpen}
-        onClose={handleProfileEditClose}
-        address={effectiveAddress}
-        initialBio={bioText}
-        initialSection={editInitialSection}
-        onBuyChainName={openClaimChainName}
-      />
-      <ClaimChainNameModal
-        open={claimChainNameOpen}
-        onClose={() => setClaimChainNameOpen(false)}
-        address={effectiveAddress}
-      />
-    </>
+    <ProfileEditModal
+      open={editOpen}
+      onClose={handleProfileEditClose}
+      address={effectiveAddress}
+      initialBio={bioText}
+      initialSection={editInitialSection}
+    />
   );
 
   return standalone ? (
